Move UTCDateTime class from opengever.activity.model to opengever.base.model.
